moved reverse lookup tests into the lookup tests file
